Made resource page mobile friendly
