+++ src/components/ResourcePage/LeftPanel.css
@@ -2,10 +2,6 @@
   margin-right: 50px;
 }
 
-.title {
-  font-weight: bold;
-}
-
 .map-header {
   display: flex;
   justify-content: space-between;
