+++ src/components/ResourcePage/ResourcePage.js
@@ -1,12 +1,53 @@
 import React from 'react';
 import { withRouter } from "react-router-dom";
+import Container from 'react-bootstrap/Container';
+import Row from 'react-bootstrap/Row';
+import Col from 'react-bootstrap/Col';
+import Modal from 'react-bootstrap/Modal';
+import Button from 'react-bootstrap/Button';
+import Form from 'react-bootstrap/Form';
+import PhoneIcon from '@material-ui/icons/Phone';
+import PrintIcon from '@material-ui/icons/Print';
+import WarningIcon from '@material-ui/icons/Warning';
 import { getDistance, convertDistance } from 'geolib';
+import emailjs from 'emailjs-com';
 import LeftPanel from './LeftPanel.js';
 import RightPanel from './RightPanel.js';
 import DescriptionPanel from './DescriptionPanel.js';
 
 import './ResourcePage.css';
 
+class ReportIssueModal extends React.Component {
+  render() {
+    const { submitIssue } = this.props;
+    return (
+      <Modal
+        {...this.props}
+        size="lg"
+        centered
+        dialogClassName="issue-modal"
+      >
+        <Modal.Header closeButton>
+          <Modal.Title id="contained-modal-title-vcenter">
+            Report an Issue
+          </Modal.Title>
+        </Modal.Header>
+        <Modal.Body>
+          <Form onSubmit={submitIssue}>
+            <Form.Group controlId="formBasicTextarea">
+              <Form.Control
+                size="lg"
+                as="textarea"
+                required
+              />
+            </Form.Group>
+            <Button type="submit">Submit</Button>
+          </Form>
+        </Modal.Body>
+      </Modal>
+    );
+  }
+}
 
 class ResourcePage extends React.Component {
   constructor(props) {
@@ -22,6 +63,7 @@ class ResourcePage extends React.Component {
       targetLong: null,
       tags: [],
       searchError: null,
+      showModal: false,
     };
   }
 
@@ -46,12 +88,31 @@ class ResourcePage extends React.Component {
           location: result.location,
           services: result.services,
           distance: distanceInMiles,
-          targetLat: result.location.geo.coordinates[0],
-          targetLong: result.location.geo.coordinates[1],
+          targetLat: result.location.geo.coordinates[1],
+          targetLong: result.location.geo.coordinates[0],
           tags: result.tags,
           searchError: null,
         })
-      });
+      })
+      .catch((error) => this.setState({searchError: error}));
+  }
+
+  submitIssue = event => {
+    event.preventDefault();
+    const msg = event.target.elements.formBasicTextarea.value;
+    const templateParams = {
+      url: window.location.href,
+      message: msg,
+    }
+
+    emailjs.send(process.env.REACT_APP_EMAILJS_SERVICE_ID, process.env.REACT_APP_EMAILJS_TEMPLATE_ID, templateParams, process.env.REACT_APP_EMAILJS_USER_ID)
+      .then(result => {
+        alert('Email sent. We\'ll look into the issue.', result.text);
+      },
+      error => {
+        alert( 'An error occured. Please try again.', error.text)
+      })
+    this.setState({ showModal: false });
   }
 
   render() {
@@ -66,33 +127,63 @@ class ResourcePage extends React.Component {
       targetLong,
       tags,
       searchError,
+      showModal,
     } = this.state;
 
     return (
       <div className="ResourcePage">
+        <ReportIssueModal
+          show={showModal}
+          onSubmit={this.submitIssue}
+          onHide={() => this.setState({ showModal: false})}
+        />
         {searchError &&
           <p style={{'textAlign': 'center'}}>Something unexpected happened. <a href="/">Click here</a> to return to the homepage.</p>
         }
         {!searchError && targetLat && targetLong &&
-          <div className="resource-page-container">
-            <div className="panel-container">
-              <LeftPanel
-                name={name}
-                location={location}
-                distance={distance}
-                targetLat={targetLat}
-                targetLong={targetLong}
-              />
-              <RightPanel
-                phone={phone}
-                description={description}
-                services={services}
-              />
-            </div>
-            <DescriptionPanel
-              tags={tags}
-            />
-          </div>
+          <Container fluid style={{height: '100%'}}>
+            <Row>
+              <Col xs={12} sm={12} md={6} lg={6} style={{paddingLeft: '0px', paddingRight: '0px'}}>
+                <h1 className="title">{name}</h1>
+              </Col>
+              <Col xs={12} sm={12} md={6} lg={6} style={{paddingLeft: '0px', paddingRight: '0px'}}>
+                <div className="right-icons">
+                  <span className="phone">
+                    <PhoneIcon/> {phone}
+                  </span> 
+                  <span className="print" onClick={() => window.print()}>
+                    <PrintIcon/> Print Page
+                  </span>
+                  <span className="error" onClick={() => this.setState({ showModal: true })}>
+                    <WarningIcon/> Report an Issue
+                  </span>
+                </div>
+              </Col>
+              <Col xs={12} sm={12} md={6} lg={6} style={{paddingLeft: '0px', paddingRight: '0px'}}>
+                <LeftPanel
+                  name={name}
+                  location={location}
+                  distance={distance}
+                  targetLat={targetLat}
+                  targetLong={targetLong}
+                />
+              </Col>
+              <Col xs={12} sm={12} md={6} lg={6} style={{paddingLeft: '0px', paddingRight: '0px'}}>
+                <RightPanel
+                  phone={phone}
+                  description={description}
+                  services={services}
+                />
+              </Col>
+            </Row>
+            <Row>
+              <Col xs={12} sm={12} md={6} lg={6} style={{paddingLeft: '0px', paddingRight: '0px'}}>
+                <DescriptionPanel
+                  tags={tags}
+                />
+              </Col>
+            </Row>
+          </Container>
         }
       </div>
     );
