+++ src/components/ResourcePage/RightPanel.css
@@ -3,37 +3,6 @@
   flex-direction: column;
 }
 
-.right-icons {
-  font-weight: bold;
-  display: flex;
-  flex-direction: column;
-  margin-left: auto;
-}
-
-.print:hover {
-  cursor: pointer;
-}
-
-.error:hover {
-  cursor: pointer;
-}
-
 .summary-title {
   font-weight: bold;
-}
-
-.issue-modal .modal-body {
-  padding: 0px 1rem !important;
-}
-
-.issue-modal .modal-content .modal-body form {
-  padding: 0px !important;
-}
-
-.issue-modal .form-group {
-  margin-bottom: 0px !important;
-}
-
-.issue-modal .modal-content .modal-body .btn {
-  float: right;
 }
